REG-5021: Fikset inntekt slik at den nå henter syntetiske meldinger og legger disse inn i inntektstub

+++ src/test/java/no/nav/registre/inntekt/testUtils/InntektGenerator.java
@@ -5,7 +5,7 @@ import no.nav.registre.inntekt.domain.RsInntektsinformasjonsType;
 
 public class InntektGenerator {
 
-    public static RsInntekt genererInntekt(int beloep) {
+    public static RsInntekt genererInntekt(double beloep) {
         RsInntekt inntektEntry = new RsInntekt();
         inntektEntry.setAar("2019");
         inntektEntry.setBeloep(beloep);
